Use storyboard for UI test host app.
